feat(pie-models): Refactored Feedback configurations.
